fixed circular dependencies and type issues
