+++ examples/with-sdk-js/src/app/page.tsx
@@ -16,7 +16,7 @@ import {
   useModal,
   useTurnkey,
 } from "@turnkey/react-wallet-kit";
-import { SessionKey } from "@turnkey/sdk-js/dist/__storage__/base";
+import { SessionKey } from "@turnkey/react-wallet-kit";
 import { WalletType } from "@turnkey/wallet-stamper";
 import { ExportType } from "@turnkey/react-wallet-kit/dist/components/export";
