+++ packages/react-wallet-kit/src/components/auth/Phone.tsx
@@ -3,7 +3,7 @@ import { ActionButton, IconButton } from "../design/Buttons";
 import { PhoneInputBox } from "../design/Inputs";
 import { faArrowRight } from "@fortawesome/free-solid-svg-icons";
 import clsx from "clsx";
-import { useTurnkey } from "../../providers";
+import { useTurnkey } from "../../providers/client/Hook";
 
 interface PhoneNumberInputProps {
   onContinue?: (phone: string, formattedPhone: string) => void;
