+++ packages/react-wallet-kit/src/components/auth/Wallet.tsx
@@ -1,6 +1,6 @@
 import { WalletProvider, WalletType } from "@turnkey/wallet-stamper";
 import { ActionButton } from "../design/Buttons";
-import { useModal } from "../../providers";
+import { useModal } from "../../providers/modal/Hook";
 import { EthereumLogo, SolanaLogo } from "../design/Svg";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faChevronRight, faClose } from "@fortawesome/free-solid-svg-icons";
