+++ packages/react-wallet-kit/src/components/user/RemoveOAuthProvider.tsx
@@ -2,7 +2,8 @@ import { faTriangleExclamation } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { ActionButton } from "../design/Buttons";
 import { useState } from "react";
-import { useModal, useTurnkey } from "../../providers";
+import { useModal } from "../../providers/modal/Hook";
+import { useTurnkey } from "../../providers/client/Hook";
 import { TurnkeyError, TurnkeyErrorCodes } from "@turnkey/sdk-types";
 import clsx from "clsx";
