+++ packages/react-wallet-kit/src/components/user/UpdatePhoneNumber.tsx
@@ -1,6 +1,7 @@
 import { faPhone } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { useModal, useTurnkey } from "../../providers";
+import { useModal } from "../../providers/modal/Hook";
+import { useTurnkey } from "../../providers/client/Hook";
 import { ActionButton } from "../design/Buttons";
 import { useState } from "react";
 import { PhoneInputBox } from "../design/Inputs";
