+++ packages/react-wallet-kit/src/providers/modal/Hook.tsx
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { ModalContext, ModalContextType } from "./Provider";
+
+export const useModal = (): ModalContextType => {
+  const context = useContext(ModalContext);
+  if (!context) throw new Error("useModal must be used within ModalProvider");
+  return context;
+};
