+++ packages/sdk-browser/src/__inputs__/public_api.types.ts
@@ -44,6 +44,10 @@ export type paths = {
     /** Get details about a Policy */
     post: operations["PublicApiService_GetPolicy"];
   };
+  "/public/v1/query/get_policy_evaluations": {
+    /** Get the policy evaluations for an Activity */
+    post: operations["PublicApiService_GetPolicyEvaluations"];
+  };
   "/public/v1/query/get_private_key": {
     /** Get details about a Private Key */
     post: operations["PublicApiService_GetPrivateKey"];
@@ -457,6 +461,28 @@ export type definitions = {
     createdAt: definitions["externaldatav1Timestamp"];
     updatedAt: definitions["externaldatav1Timestamp"];
   };
+  externalactivityv1PolicyEvaluation: {
+    /** @description Unique identifier for a given policy evaluation. */
+    id: string;
+    /** @description Unique identifier for a given Activity. */
+    activityId: string;
+    /** @description Unique identifier for the Organization the Activity belongs to. */
+    organizationId: string;
+    /** @description Unique identifier for the Vote associated with this policy evaluation. */
+    voteId: string;
+    /** @description Detailed evaluation result for each Policy that was run. */
+    policyEvaluations: definitions["privateumpv1PolicyEvaluation"][];
+    createdAt: definitions["externaldatav1Timestamp"];
+  };
+  externalactivityv1UpdateProxyAuthConfigRequest: {
+    /** @enum {string} */
+    type: "ACTIVITY_TYPE_UPDATE_PROXY_AUTH_CONFIG";
+    /** @description Timestamp (in milliseconds) of the request, used to verify liveness of user requests. */
+    timestampMs: string;
+    /** @description Unique identifier for a given Organization. */
+    organizationId: string;
+    parameters: definitions["v1UpdateProxyAuthConfigIntent"];
+  };
   externaldatav1Address: {
     format?: definitions["v1AddressFormat"];
     address?: string;
@@ -475,6 +501,24 @@ export type definitions = {
     /** @description Unique identifiers of quorum set members. */
     userIds: string[];
   };
+  externaldatav1SmartContractInterface: {
+    /** @description The Organization the Smart Contract Interface belongs to. */
+    organizationId: string;
+    /** @description Unique identifier for a given Smart Contract Interface (ABI or IDL). */
+    smartContractInterfaceId: string;
+    /** @description The address corresponding to the Smart Contract or Program. */
+    smartContractAddress: string;
+    /** @description The JSON corresponding to the Smart Contract Interface (ABI or IDL). */
+    smartContractInterface: string;
+    /** @description The type corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
+    type: string;
+    /** @description The label corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
+    label: string;
+    /** @description The notes corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
+    notes: string;
+    createdAt: definitions["externaldatav1Timestamp"];
+    updatedAt: definitions["externaldatav1Timestamp"];
+  };
   externaldatav1Timestamp: {
     seconds: string;
     nanos: string;
@@ -482,6 +526,10 @@ export type definitions = {
   immutableactivityv1Address: {
     format?: definitions["v1AddressFormat"];
     address?: string;
+  };
+  privateumpv1PolicyEvaluation: {
+    policyId?: string;
+    outcome?: definitions["v1Outcome"];
   };
   protobufAny: {
     "@type"?: string;
@@ -1633,12 +1681,7 @@ export type definitions = {
     /** @description Unique identifier for a given Private Key. */
     privateKeyId: string;
   };
-  v1DisableUserInitiatedAuthIntent: {
-    /** @description Unique identifier for a given User. (representing the turnkey signer user id) */
-    userId?: string;
-    /** @description Unique identifier for a given Policy. (representing the turnkey signer associated policy) */
-    policyId?: string;
-  };
+  v1DisableUserInitiatedAuthIntent: { [key: string]: unknown };
   v1DisableUserInitiatedAuthRequest: {
     /** @enum {string} */
     type: "ACTIVITY_TYPE_DISABLE_USER_INITIATED_AUTH";
@@ -1729,20 +1772,8 @@ export type definitions = {
     parameters: definitions["v1EnableUserInitiatedAuthIntent"];
   };
   v1EnableUserInitiatedAuthResult: {
-    /** @description A User ID. */
+    /** @description A User ID with permission to initiate authentication. */
     userId: string;
-    /** @description A Policy ID. */
-    policyId: string;
-  };
-  v1EwkSettingsParams: {
-    /** @description Enable Sign in with Apple */
-    appleEnabled?: boolean;
-    /** @description Enable Sign in with Google */
-    googleEnabled?: boolean;
-    /** @description Enable Sign in with Facebook */
-    facebookEnabled?: boolean;
-    /** @description Whether to open OAuth providers in-page instead of a popup */
-    openOauthInPage?: boolean;
   };
   v1ExportPrivateKeyIntent: {
     /** @description Unique identifier for a given Private Key. */
@@ -1821,7 +1852,8 @@ export type definitions = {
     | "FEATURE_NAME_EMAIL_RECOVERY"
     | "FEATURE_NAME_WEBHOOK"
     | "FEATURE_NAME_SMS_AUTH"
-    | "FEATURE_NAME_OTP_EMAIL_AUTH";
+    | "FEATURE_NAME_OTP_EMAIL_AUTH"
+    | "FEATURE_NAME_AUTH_KIT";
   /** @enum {string} */
   v1FiatOnRampBlockchainNetwork:
     | "FIAT_ON_RAMP_BLOCKCHAIN_NETWORK_BITCOIN"
@@ -1996,6 +2028,15 @@ export type definitions = {
     /** @description A list of Policies. */
     policies: definitions["v1Policy"][];
   };
+  v1GetPolicyEvaluationsRequest: {
+    /** @description Unique identifier for a given Organization. */
+    organizationId: string;
+    /** @description Unique identifier for a given Activity. */
+    activityId: string;
+  };
+  v1GetPolicyEvaluationsResponse: {
+    policyEvaluations: definitions["externalactivityv1PolicyEvaluation"][];
+  };
   v1GetPolicyRequest: {
     /** @description Unique identifier for a given Organization. */
     organizationId: string;
@@ -2040,7 +2081,7 @@ export type definitions = {
   };
   v1GetSmartContractInterfaceResponse: {
     /** @description Object to be used in conjunction with Policies to guard transaction signing. */
-    smartContractInterface: definitions["v1SmartContractInterface"];
+    smartContractInterface: definitions["externaldatav1SmartContractInterface"];
   };
   v1GetSmartContractInterfacesRequest: {
     /** @description Unique identifier for a given Organization. */
@@ -2048,12 +2089,12 @@ export type definitions = {
   };
   v1GetSmartContractInterfacesResponse: {
     /** @description A list of Smart Contract Interfaces. */
-    smartContractInterfaces: definitions["v1SmartContractInterface"][];
+    smartContractInterfaces: definitions["externaldatav1SmartContractInterface"][];
   };
   v1GetSubOrgIdsRequest: {
     /** @description Unique identifier for the parent Organization. This is used to find sub-organizations within it. */
     organizationId: string;
-    /** @description Specifies the type of filter to apply, i.e 'CREDENTIAL_ID', 'NAME', 'USERNAME', 'EMAIL', 'PHONE_NUMBER', 'OIDC_TOKEN' or 'PUBLIC_KEY' */
+    /** @description Specifies the type of filter to apply, i.e 'CREDENTIAL_ID', 'NAME', 'USERNAME', 'EMAIL', 'PHONE_NUMBER', 'OIDC_TOKEN', 'WALLET_ACCOUNT_ADDRESS' or 'PUBLIC_KEY' */
     filterType?: string;
     /** @description The value of the filter to apply for the specified type. For example, a specific email or name string. */
     filterValue?: string;
@@ -2722,6 +2763,14 @@ export type definitions = {
     /** @description Signed JWT containing an expiry, public key, session type, user id, and organization id */
     session: string;
   };
+  /** @enum {string} */
+  v1Outcome:
+    | "OUTCOME_ALLOW"
+    | "OUTCOME_DENY_EXPLICIT"
+    | "OUTCOME_DENY_IMPLICIT"
+    | "OUTCOME_REQUIRES_CONSENSUS"
+    | "OUTCOME_REJECTED"
+    | "OUTCOME_ERROR";
   v1Pagination: {
     /** @description A limit of the number of object to be returned, between 1 and 100. Defaults to 10. */
     limit?: string;
@@ -2804,13 +2853,7 @@ export type definitions = {
     /** Format: int32 */
     verificationTokenExpirationSeconds?: number;
     /** Format: int32 */
-    otpSessionExpirationSeconds?: number;
-    /** Format: int32 */
-    oauthSessionExpirationSeconds?: number;
-    /** Format: int32 */
-    passkeySessionExpirationSeconds?: number;
-    /** Format: int32 */
-    walletSessionExpirationSeconds?: number;
+    sessionExpirationSeconds?: number;
     /** Format: date-time */
     createdAt?: string;
     /** Format: date-time */
@@ -2818,10 +2861,8 @@ export type definitions = {
     otpAlphanumeric?: boolean;
     /** Format: int32 */
     otpLength?: number;
-    socialLinking?: boolean;
-    policyId?: string;
     proxyId?: string;
-    ewkSettings?: string;
+    walletKitSettings?: string;
   };
   v1PublicKeyCredentialWithAttestation: {
     id: string;
@@ -3133,24 +3174,6 @@ export type definitions = {
     appidExclude?: boolean;
     credProps?: definitions["v1CredPropsAuthenticationExtensionsClientOutputs"];
   };
-  v1SmartContractInterface: {
-    /** @description The Organization the Smart Contract Interface belongs to. */
-    organizationId: string;
-    /** @description Unique identifier for a given Smart Contract Interface (ABI or IDL). */
-    smartContractInterfaceId: string;
-    /** @description The address corresponding to the Smart Contract or Program. */
-    smartContractAddress: string;
-    /** @description The JSON corresponding to the Smart Contract Interface (ABI or IDL). */
-    smartContractInterface: string;
-    /** @description The type corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
-    type: string;
-    /** @description The label corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
-    label: string;
-    /** @description The notes corresponding to the Smart Contract Interface (either ETHEREUM or SOLANA). */
-    notes: string;
-    createdAt: definitions["externaldatav1Timestamp"];
-    updatedAt: definitions["externaldatav1Timestamp"];
-  };
   v1SmartContractInterfaceReference: {
     smartContractInterfaceId?: string;
     smartContractAddress?: string;
@@ -3280,7 +3303,7 @@ export type definitions = {
   v1UpdateProxyAuthConfigIntent: {
     /** @description Updated list of allowed origins for CORS. */
     allowedOrigins?: string[];
-    /** @description Updated list of allowed origins for CORS. */
+    /** @description Updated list of allowed proxy authentication methods. */
     allowedAuthMethods?: string[];
     /** @description Custom 'from' address for auth-related emails. */
     sendFromEmailAddress?: string;
@@ -3294,8 +3317,8 @@ export type definitions = {
     emailCustomizationParams?: definitions["v1EmailCustomizationParams"];
     /** @description Overrides for auth-related SMS content. */
     smsCustomizationParams?: definitions["v1SmsCustomizationParams"];
-    /** @description Overrides for EWK related settings. */
-    ewkSettings?: definitions["v1EwkSettingsParams"];
+    /** @description Overrides for react wallet kit related settings. */
+    walletKitSettings?: definitions["v1WalletKitSettingsParams"];
     /**
      * Format: int32
      * @description OTP code lifetime in seconds.
@@ -3308,24 +3331,9 @@ export type definitions = {
     verificationTokenExpirationSeconds?: number;
     /**
      * Format: int32
-     * @description OTP session lifetime in seconds.
+     * @description Session lifetime in seconds.
      */
-    otpSessionExpirationSeconds?: number;
-    /**
-     * Format: int32
-     * @description Passkey session lifetime in seconds.
-     */
-    passkeySessionExpirationSeconds?: number;
-    /**
-     * Format: int32
-     * @description Wallet session lifetime in seconds.
-     */
-    walletSessionExpirationSeconds?: number;
-    /**
-     * Format: int32
-     * @description OAuth session lifetime in seconds.
-     */
-    oauthSessionExpirationSeconds?: number;
+    sessionExpirationSeconds?: number;
     /** @description Enable alphanumeric OTP codes. */
     otpAlphanumeric?: boolean;
     /**
@@ -3333,19 +3341,11 @@ export type definitions = {
      * @description Desired OTP code length (6–9).
      */
     otpLength?: number;
-    /** @description Enable social linking (userEmail <-> gmail) */
-    socialLinking?: boolean;
   };
-  v1UpdateProxyAuthConfigRequest: {
-    /** @enum {string} */
-    type: "ACTIVITY_TYPE_UPDATE_PROXY_AUTH_CONFIG";
-    /** @description Timestamp (in milliseconds) of the request, used to verify liveness of user requests. */
-    timestampMs: string;
-    /** @description Unique identifier for a given Organization. */
-    organizationId: string;
-    parameters: definitions["v1UpdateProxyAuthConfigIntent"];
+  v1UpdateProxyAuthConfigResult: {
+    /** @description Unique identifier for a given User. (representing the turnkey signer user id) */
+    configId?: string;
   };
-  v1UpdateProxyAuthConfigResult: { [key: string]: unknown };
   v1UpdateRootQuorumIntent: {
     /**
      * Format: int32
@@ -3641,6 +3641,16 @@ export type definitions = {
     /** @description Address format used to generate a wallet Acccount. */
     addressFormat: definitions["v1AddressFormat"];
   };
+  v1WalletKitSettingsParams: {
+    /** @description Enable Sign in with Apple */
+    appleEnabled?: boolean;
+    /** @description Enable Sign in with Google */
+    googleEnabled?: boolean;
+    /** @description Enable Sign in with Facebook */
+    facebookEnabled?: boolean;
+    /** @description Whether to open OAuth providers in-page instead of a popup */
+    openOauthInPage?: boolean;
+  };
   v1WalletParams: {
     /** @description Human-readable name for a Wallet. */
     walletName: string;
@@ -3843,6 +3853,24 @@ export type operations = {
       /** A successful response. */
       200: {
         schema: definitions["v1GetPolicyResponse"];
+      };
+      /** An unexpected error response. */
+      default: {
+        schema: definitions["rpcStatus"];
+      };
+    };
+  };
+  /** Get the policy evaluations for an Activity */
+  PublicApiService_GetPolicyEvaluations: {
+    parameters: {
+      body: {
+        body: definitions["v1GetPolicyEvaluationsRequest"];
+      };
+    };
+    responses: {
+      /** A successful response. */
+      200: {
+        schema: definitions["v1GetPolicyEvaluationsResponse"];
       };
       /** An unexpected error response. */
       default: {
@@ -5222,7 +5250,7 @@ export type operations = {
   PublicApiService_UpdateProxyAuthConfig: {
     parameters: {
       body: {
-        body: definitions["v1UpdateProxyAuthConfigRequest"];
+        body: definitions["externalactivityv1UpdateProxyAuthConfigRequest"];
       };
     };
     responses: {
