+++ packages/sdk-types/src/__inputs__/auth_proxy.swagger.json
@@ -15,7 +15,7 @@
     "/v1/account": {
       "post": {
         "summary": "Get Account",
-        "description": "Return sub-organization ID associated with a given contact or public identifier.",
+        "description": "Return account ID associated with a given phone number, email, public key, credential ID or OIDC token.",
         "operationId": "AuthProxyService_GetAccount",
         "responses": {
           "200": {
@@ -111,7 +111,7 @@
     "/v1/otp_login": {
       "post": {
         "summary": "OTP Login",
-        "description": "Login using a verified OTP token and public key.",
+        "description": "Login using a verification token and public key.",
         "operationId": "AuthProxyService_OtpLogin",
         "responses": {
           "200": {
@@ -143,7 +143,7 @@
     "/v1/otp_verify": {
       "post": {
         "summary": "Verify OTP",
-        "description": "Verify the OTP code previously sent to the user's contact.",
+        "description": "Verify the OTP code previously sent to the user's contact and return a verification token.",
         "operationId": "AuthProxyService_VerifyOtp",
         "responses": {
           "200": {
@@ -175,7 +175,7 @@
     "/v1/signup": {
       "post": {
         "summary": "Signup",
-        "description": "Create a new sub-organization and user via the proxy API.",
+        "description": "Onboard a a new user.",
         "operationId": "AuthProxyService_Signup",
         "responses": {
           "200": {
@@ -233,7 +233,7 @@
             }
           }
         ],
-        "tags": ["Organizations"]
+        "tags": ["AuthKit"]
       }
     }
   },
@@ -452,10 +452,7 @@
         "openOAuthInPage": {
           "type": "boolean"
         },
-        "passkeySessionExpirationSeconds": {
-          "type": "string"
-        },
-        "walletSessionExpirationSeconds": {
+        "sessionExpirationSeconds": {
           "type": "string"
         },
         "organizationId": {
@@ -471,8 +468,7 @@
         "passkeyEnabled",
         "walletEnabled",
         "openOAuthInPage",
-        "passkeySessionExpirationSeconds",
-        "walletSessionExpirationSeconds",
+        "sessionExpirationSeconds",
         "organizationId"
       ]
     },
